requirements.txt
torch
transformers
peft
opencv-python
pillow
numpy
tqdm
wandb
decord